pyproject.toml
[project]
name = "arena-gaming-bot"
version = "0.1.0"
description = "Bot Discord moderne pour League of Legends et Teamfight Tactics"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.4.0",
    "aiohttp>=3.11.10",
    "python-dotenv>=1.0.1",
    "pillow>=11.0.0",
    "requests>=2.32.3",
    "watchdog>=6.0.0",
    "pydantic>=2.10.2",
    "structlog>=24.4.0",
    "ratelimiter>=1.2.0",
]
